// src/pages/gym-app/guest/payment-receipt/type.ts
export type I_PAYMENT_RECEIPT = {
  success: boolean;
  message: string;
  data: {
    _id: string;
    memberId: {
      _id: string;
      generatedId: string;
      userId: string;
      memberName: string;
      address: string;
      mobile: string;
    };

    subscriptionId: [
      {
        subscriptionId: string;
        paidAmount: number;
        dueAmount: number;
        startDate: string;
        endDate: string;
        amount: number;
        planName: string;
        month: string;
        paidStatus: number;
        prevDueAmount: number;
      }
    ];
  

    receiptNo: string;
    receiptDate: string;
    totalPaidAmount: number;
    totalDueAmount: number;
    totalPrevDueAmount: number;
    paymentMode: string;
    paymentStatus: string;
    status: number;
    createdAt: string;
    updatedAt: string;
  };
  userDetails: {
    _id: string,
    fullName: string ,
    gymName: string,
    roleId: string,
    mobile: number,
    email: string,
    address: string,
    status: 1,
    fullImgUrl: string,
}
};